+++ h2integrate/core/supported_models.py
@@ -18,6 +18,7 @@ from h2integrate.finances.numpy_financial_npv import NumpyFinancialNPV
 from h2integrate.resource.wind.openmeteo_wind import OpenMeteoHistoricalWindResource
 from h2integrate.storage.generic_storage_cost import GenericStorageCostModel
 from h2integrate.storage.hydrogen.mch_storage import MCHTOLStorageCostModel
+from h2integrate.converters.steel.cmu_eaf_cost import CMUElectricArcFurnaceCostModel
 from h2integrate.converters.wind.atb_wind_cost import ATBWindPlantCostModel
 from h2integrate.storage.battery.pysam_battery import PySAMBatteryPerformanceModel
 from h2integrate.transporters.generic_combiner import GenericCombinerPerformanceModel
@@ -117,6 +118,12 @@ from h2integrate.resource.solar.nlr_developer_goes_api_models import (
     GOESFullDiscSolarAPI,
     GOESAggregatedSolarAPI,
 )
+from h2integrate.converters.steel.cmu_electric_arc_furnace_dri import (
+    CMUElectricArcFurnaceDRIPerformanceComponent,
+)
+from h2integrate.converters.steel.cmu_electric_arc_furnace_scrap import (
+    CMUElectricArcFurnaceScrapOnlyPerformanceComponent,
+)
 from h2integrate.converters.water_power.hydro_plant_run_of_river import (
     RunOfRiverHydroCostModel,
     RunOfRiverHydroPerformanceModel,
@@ -162,6 +169,9 @@ from h2integrate.control.control_strategies.storage.optimized_pyomo_controller i
 )
 from h2integrate.control.control_strategies.storage.simple_openloop_controller import (
     SimpleStorageOpenLoopController,
+)
+from h2integrate.control.control_strategies.storage.plm_openloop_storage_controller import (
+    PeakLoadManagementHeuristicOpenLoopStorageController,
 )
 from h2integrate.control.control_rules.storage.pyomo_storage_rule_min_operating_cost import (
     PyomoRuleStorageMinOperatingCosts,
@@ -227,6 +237,11 @@ supported_models = {
     "NaturalGasEAFPlantCostComponent": NaturalGasEAFPlantCostComponent,  # standalone model
     "HydrogenEAFPlantPerformanceComponent": HydrogenEAFPlantPerformanceComponent,
     "HydrogenEAFPlantCostComponent": HydrogenEAFPlantCostComponent,  # standalone model
+    "CMUElectricArcFurnaceScrapOnlyPerformanceComponent": (
+        CMUElectricArcFurnaceScrapOnlyPerformanceComponent
+    ),
+    "CMUElectricArcFurnaceDRIPerformanceComponent": CMUElectricArcFurnaceDRIPerformanceComponent,
+    "CMUElectricArcFurnaceCostModel": CMUElectricArcFurnaceCostModel,
     "ReverseOsmosisPerformanceModel": ReverseOsmosisPerformanceModel,
     "ReverseOsmosisCostModel": ReverseOsmosisCostModel,
     "SimpleAmmoniaPerformanceModel": SimpleAmmoniaPerformanceModel,
@@ -278,6 +293,9 @@ supported_models = {
     # Control
     "SimpleStorageOpenLoopController": SimpleStorageOpenLoopController,
     "DemandOpenLoopStorageController": DemandOpenLoopStorageController,
+    "PeakLoadManagementHeuristicOpenLoopStorageController": (
+        PeakLoadManagementHeuristicOpenLoopStorageController
+    ),
     "HeuristicLoadFollowingStorageController": HeuristicLoadFollowingStorageController,
     "OptimizedDispatchStorageController": OptimizedDispatchStorageController,
     "GenericDemandComponent": GenericDemandComponent,